requirements.txt
sqlalchemy == 2.0.28
Jinja2 == 3.1.2
Flask == 3.0.1
mysql-connector-python==8.2.0
mysql-connector==2.2.9
mysqlclient==2.1.0